% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_setup.R
\name{f_setup}
\alias{f_setup}
\title{Private}
\usage{
f_setup()
}
\value{
Setting up the global environment
}
\description{
The function \code{f_setup} Private
}
\details{
Loads installed packages, custom functions, and transaction cost tables into the global environment
}
\examples{
require(rpackagemock1)
f_setup()

}

